update 1-strncat.c file

+++ pointers_arrays_strings/1-strncat.c
@@ -1,5 +1,4 @@
 #include "main.h"
-#include <stddef.h>
 
 /**
 * *_strncat - concatenates two strings.
@@ -13,7 +12,7 @@
 
 char *_strncat(char *dest, char *src, int n)
 {
-	size_t i = 0, j = 0;
+	int i = 0, j = 0;
 
 	while (dest[i] != '\0')
 		{
